Refactor: Split components for the status timeline component
