+++ react/admin/ReturnDetails/ReturnDetailsContainer.tsx
@@ -11,7 +11,7 @@ import { ContactDetails } from './components/ContactDetails'
 import { PickupAddress } from './components/PickupAddress'
 import { AdminLoader } from '../AdminLoader'
 import { ReturnValues } from './components/ReturnValues/ReturnValues'
-import { StatusTimeline } from './components/StatusTimeline'
+import { StatusTimeline } from './components/StatusTimeline/StatusTimeline'
 
 type Pages = 'return-details' | 'verify-items'
