+++ react/styles.global.css
@@ -5,8 +5,8 @@
 }
 
 .status-icon-not-visited {
-  width: 26px;
-  height: 26px;
+  width: 30px;
+  height: 30px;
   border: 3px solid #134cd8;
   border-radius: 50%;
   display: inline-block;
@@ -15,7 +15,7 @@
   padding-left: 3px;
   background: #fff;
   margin-right: 20px;
-  margin-left: 5px;
+  margin-left: 3px;
 }
 
 .status-icon-checked {
